Created a GeocodingModelPoller and ImportModelPoller with specific configuration for each type of model
