+++ lib/assets/javascripts/cartodb/common/background_polling/models/geocoding_model.js
@@ -1,7 +1,6 @@
 var Backbone = require('backbone');
 var _ = require('underscore');
-var Poller = require('./poller');
-var POLLING_INTERVAL = 2000;
+var GeocodingModelPoller = require('./geocoding_model_poller');
 
 /**
  *  Geocoding model
@@ -38,15 +37,7 @@ module.exports = cdb.core.Model.extend({
     var self = this;
     this._initBinds();
     _.extend(this.options, opts);
-    this.poller = new Poller(this, {
-      interval: POLLING_INTERVAL,
-      condition: function(model) {
-        return model.hasFailed() || model.hasCompleted();
-      },
-      error: function() {
-        self.trigger("change");
-      }
-    });
+    this.poller = new GeocodingModelPoller(this);
     if (this.options.startPollingAutomatically) {
       this._checkModel();
     };
