+++ lib/assets/javascripts/cartodb/common/background_polling/models/import_model.js
@@ -1,9 +1,6 @@
 var cdb = require('cartodb.js');
 var cdbAdmin = require('cdb.admin');
-var Poller = require('./poller');
-var POLLING_INTERVAL = 2000; // Interval time between poll checkings
-var POLLING_INTERVAL_MULTIPLIER = 2.5;  // Multiply interval by this number
-var POLLING_REQUESTS_BEFORE_INTERVAL_CHANGE = 30; // Max tries until interval change
+var ImportModelPoller = require('./import_model_poller');
 
 /**
  *  New import model that controls
@@ -25,18 +22,7 @@ module.exports = cdb.core.Model.extend({
 
   initialize: function() {
     this._initBinds();
-    this.poller = new Poller(this, {
-      interval: function(numberOfRequests) {
-        if (numberOfRequests >= POLLING_REQUESTS_BEFORE_INTERVAL_CHANGE) {
-          return POLLING_INTERVAL * POLLING_INTERVAL_MULTIPLIER;
-        }
-        return POLLING_INTERVAL;
-      },
-      condition: function(model) {
-        var state = model.get('state');
-        return (state === "complete" || state === "failure");
-      }
-    });
+    this.poller = new ImportModelPoller(this);
   },
 
   _initBinds: function() {
