+++ lib/assets/test/spec/cartodb/common/background_importer/geocoding_model.spec.js
@@ -48,19 +48,6 @@ describe('common/background_polling/geocoding_model', function() {
 
       expect(this.model.poller.start).toHaveBeenCalled();
     });
-
-    it('should trigger a change event if an error occurs', function() {
-      var onErrorCallback = jasmine.createSpy('error');
-
-      this.model.pollCheck();
-      this.model.bind('change', onErrorCallback);
-
-      jasmine.clock().tick(2001);
-
-      this.model.fetch.calls.mostRecent().args[0].error();
-
-      expect(onErrorCallback).toHaveBeenCalled();
-    });
   });
 
   describe('.destroyCheck', function() {
@@ -81,41 +68,6 @@ describe('common/background_polling/geocoding_model', function() {
       this.model.destroyCheck();
 
       expect(this.model.poller.stop).toHaveBeenCalled();
-    });
-  });
-
-  describe('polling', function() {
-
-    beforeEach(function() {
-      jasmine.clock().install();
-      this.model.set('id', '1');
-      this.model.sync = function(a,b,opts) {
-        opts.success();
-      };
-
-      spyOn(this.model, 'fetch').and.callThrough();
-    });
-
-    afterEach(function() {
-      jasmine.clock().uninstall();
-    });
-
-    it('should stop polling when geocoding has failed', function() {
-      this.model.set('state', 'failed');
-      this.model.pollCheck();
-
-      jasmine.clock().tick(4000);
-
-      expect(this.model.fetch.calls.count()).toBe(1);
-    });
-
-    it('should stop polling when geocoding has completed', function() {
-      this.model.set('state', 'finished');
-      this.model.pollCheck();
-
-      jasmine.clock().tick(4000);
-
-      expect(this.model.fetch.calls.count()).toBe(1);
     });
   });
