+++ lib/assets/test/spec/cartodb/common/background_importer/imports_model.spec.js
@@ -219,58 +219,5 @@ describe('common/background_polling/imports_model', function() {
       this.imp.set('item_queue_id', 'vamos-neno');
       expect(this.imp.pollCheck).toHaveBeenCalled();
     });
-
-    describe('polling', function() {
-
-      beforeEach(function() {
-        jasmine.clock().install();
-        this.model = this.imp;
-        this.model.set('id', '1');
-        this.model.sync = function(a,b,opts) {
-          opts.success();
-        };
-
-        spyOn(this.model, 'fetch').and.callThrough();
-      });
-
-      afterEach(function() {
-        jasmine.clock().uninstall();
-      });
-
-      it('should stop polling when geocoding has failed', function() {
-        this.model.set('state', 'failure');
-        this.model.pollCheck();
-
-        jasmine.clock().tick(4000);
-
-        expect(this.model.fetch.calls.count()).toBe(1);
-      });
-
-      it('should stop polling when geocoding has completed', function() {
-        this.model.set('state', 'complete');
-        this.model.pollCheck();
-
-        jasmine.clock().tick(4000);
-
-        expect(this.model.fetch.calls.count()).toBe(1);
-      });
-
-      it('should increment the interval after a number of requests', function() {
-        this.model.pollCheck();
-
-        // 2000 is the interval period and 30 the number of requests before
-        // the interval is incremented.
-        jasmine.clock().tick(2000 * 30);
-
-        // 30 requests have been made
-        expect(this.model.fetch.calls.count()).toBe(30);
-
-        // Interval has been incremented to 2000 * 2.5
-        jasmine.clock().tick(2000 * 2.5 + 1);
-
-        // Only one more request has been made (interval has been incremented)
-        expect(this.model.fetch.calls.count()).toBe(31);
-      });
-    });
   });
 });
